requirements.txt
# Core dependencies
pyspark==3.3.2
kafka-python==2.0.2
mysql-connector-python==8.0.28

# Data processing
pandas==2.0.1
numpy==1.24.3

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0